# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flight Logger API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FlightSeatClass = typeof FlightSeatClass[keyof typeof FlightSeatClass];


export const FlightSeatClass = {
  economy: 'economy',
  premium_economy: 'premium_economy',
  business: 'business',
  first: 'first',
} as const;

export interface Flight {
  id: number;
  /** @nullable */
  flightNumber?: string | null;
  airline: string;
  /** @nullable */
  aircraftType?: string | null;
  /** IATA airport code (e.g. LHR) */
  departureAirport: string;
  /** IATA airport code (e.g. JFK) */
  arrivalAirport: string;
  /** @nullable */
  departureCity?: string | null;
  /** @nullable */
  arrivalCity?: string | null;
  departureDate: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  distanceKm?: number | null;
  seatClass: FlightSeatClass;
  /** @nullable */
  notes?: string | null;
  /**
     * Estimated CO2 in kilograms
     * @nullable
     */
  carbonKg?: number | null;
  createdAt: string;
}

export type FlightInputSeatClass = typeof FlightInputSeatClass[keyof typeof FlightInputSeatClass];


export const FlightInputSeatClass = {
  economy: 'economy',
  premium_economy: 'premium_economy',
  business: 'business',
  first: 'first',
} as const;

export interface FlightInput {
  flightNumber?: string;
  airline: string;
  aircraftType?: string;
  departureAirport: string;
  arrivalAirport: string;
  departureCity?: string;
  arrivalCity?: string;
  departureDate: string;
  durationMinutes?: number;
  distanceKm?: number;
  seatClass: FlightInputSeatClass;
  notes?: string;
}

export type FlightUpdateSeatClass = typeof FlightUpdateSeatClass[keyof typeof FlightUpdateSeatClass];


export const FlightUpdateSeatClass = {
  economy: 'economy',
  premium_economy: 'premium_economy',
  business: 'business',
  first: 'first',
} as const;

export interface FlightUpdate {
  flightNumber?: string;
  airline?: string;
  aircraftType?: string;
  departureAirport?: string;
  arrivalAirport?: string;
  departureCity?: string;
  arrivalCity?: string;
  departureDate?: string;
  durationMinutes?: number;
  distanceKm?: number;
  seatClass?: FlightUpdateSeatClass;
  notes?: string;
}

export interface Profile {
  id: number;
  name: string;
  /** @nullable */
  homeAirport?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarInitials?: string | null;
  createdAt?: string;
}

export interface ProfileInput {
  name: string;
  homeAirport?: string;
  bio?: string;
  avatarInitials?: string;
}

export type FlightStatsFlightsByClass = {[key: string]: number};

export interface FlightStats {
  totalFlights: number;
  totalDistanceKm: number;
  totalCarbonKg: number;
  uniqueAirports: number;
  uniqueCountries: number;
  uniqueAirlines: number;
  /** @nullable */
  longestFlightKm?: number | null;
  /** @nullable */
  longestFlightRoute?: string | null;
  flightsByClass: FlightStatsFlightsByClass;
}

export type CarbonStatsByYearItem = {
  year: number;
  carbonKg: number;
};

export type CarbonStatsByClassItem = {
  seatClass: string;
  carbonKg: number;
  flights: number;
};

export interface CarbonStats {
  totalCarbonKg: number;
  byYear: CarbonStatsByYearItem[];
  byClass: CarbonStatsByClassItem[];
  /** Number of trees needed to offset total carbon */
  treesEquivalent: number;
}

export interface YearlyStats {
  year: number;
  flights: number;
  distanceKm: number;
  carbonKg: number;
}

export interface DestinationStat {
  airport: string;
  /** @nullable */
  city: string | null;
  visits: number;
}

export interface AirlineStat {
  airline: string;
  flights: number;
  distanceKm: number;
}