# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Flight Logger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /flights:
    get:
      operationId: listFlights
      tags: [flights]
      summary: List all flights
      responses:
        "200":
          description: List of flights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flight"
    post:
      operationId: createFlight
      tags: [flights]
      summary: Log a new flight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlightInput"
      responses:
        "201":
          description: Flight created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flight"
  /flights/{id}:
    get:
      operationId: getFlight
      tags: [flights]
      summary: Get a flight by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Flight found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flight"
        "404":
          description: Not found
    patch:
      operationId: updateFlight
      tags: [flights]
      summary: Update a flight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlightUpdate"
      responses:
        "200":
          description: Flight updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flight"
        "404":
          description: Not found
    delete:
      operationId: deleteFlight
      tags: [flights]
      summary: Delete a flight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Flight:
      type: object
      required:
        - id
        - departureAirport
        - arrivalAirport
        - departureDate
        - airline
        - seatClass
        - createdAt
      properties:
        id:
          type: integer
        flightNumber:
          type: ["string", "null"]
        airline:
          type: string
        aircraftType:
          type: ["string", "null"]
        departureAirport:
          type: string
          description: IATA airport code (e.g. LHR)
        arrivalAirport:
          type: string
          description: IATA airport code (e.g. JFK)
        departureCity:
          type: ["string", "null"]
        arrivalCity:
          type: ["string", "null"]
        departureDate:
          type: string
          format: date
        durationMinutes:
          type: ["integer", "null"]
        distanceKm:
          type: ["integer", "null"]
        seatClass:
          type: string
          enum: [economy, premium_economy, business, first]
        notes:
          type: ["string", "null"]
        carbonKg:
          type: ["number", "null"]
          description: Estimated CO2 in kilograms
        createdAt:
          type: string
          format: date-time
    FlightInput:
      type: object
      required:
        - departureAirport
        - arrivalAirport
        - departureDate
        - airline
        - seatClass
      properties:
        flightNumber:
          type: string
        airline:
          type: string
        aircraftType:
          type: string
        departureAirport:
          type: string
        arrivalAirport:
          type: string
        departureCity:
          type: string
        arrivalCity:
          type: string
        departureDate:
          type: string
          format: date
        durationMinutes:
          type: integer
        distanceKm:
          type: integer
        seatClass:
          type: string
          enum: [economy, premium_economy, business, first]
        notes:
          type: string
    FlightUpdate:
      type: object
      properties:
        flightNumber:
          type: string
        airline:
          type: string
        aircraftType:
          type: string
        departureAirport:
          type: string
        arrivalAirport:
          type: string
        departureCity:
          type: string
        arrivalCity:
          type: string
        departureDate:
          type: string
          format: date
        durationMinutes:
          type: integer
        distanceKm:
          type: integer
        seatClass:
          type: string
          enum: [economy, premium_economy, business, first]
        notes:
          type: string
...