# lib/api-zod/src/generated/types/flight.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flight Logger API
 * OpenAPI spec version: 0.1.0
 */
import type { FlightSeatClass } from './flightSeatClass';

export interface Flight {
  id: number;
  /** @nullable */
  flightNumber?: string | null;
  airline: string;
  /** @nullable */
  aircraftType?: string | null;
  /** IATA airport code (e.g. LHR) */
  departureAirport: string;
  /** IATA airport code (e.g. JFK) */
  arrivalAirport: string;
  /** @nullable */
  departureCity?: string | null;
  /** @nullable */
  arrivalCity?: string | null;
  departureDate: Date;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  distanceKm?: number | null;
  seatClass: FlightSeatClass;
  /** @nullable */
  notes?: string | null;
  /**
     * Estimated CO2 in kilograms
     * @nullable
     */
  carbonKg?: number | null;
  createdAt: Date;
}