# lib/api-zod/src/generated/types/flightInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flight Logger API
 * OpenAPI spec version: 0.1.0
 */
import type { FlightInputSeatClass } from './flightInputSeatClass';

export interface FlightInput {
  flightNumber?: string;
  airline: string;
  aircraftType?: string;
  departureAirport: string;
  arrivalAirport: string;
  departureCity?: string;
  arrivalCity?: string;
  departureDate: Date;
  durationMinutes?: number;
  distanceKm?: number;
  seatClass: FlightInputSeatClass;
  notes?: string;
}