# lib/api-zod/src/generated/types/flightUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flight Logger API
 * OpenAPI spec version: 0.1.0
 */
import type { FlightUpdateSeatClass } from './flightUpdateSeatClass';

export interface FlightUpdate {
  flightNumber?: string;
  airline?: string;
  aircraftType?: string;
  departureAirport?: string;
  arrivalAirport?: string;
  departureCity?: string;
  arrivalCity?: string;
  departureDate?: Date;
  durationMinutes?: number;
  distanceKm?: number;
  seatClass?: FlightUpdateSeatClass;
  notes?: string;
}